pyproject.toml
[project]
name = "nexus"
version = "0.1.0"
description = "Agentic news intelligence compiler — multi-source, multi-language, always-on"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "google-genai>=1.0",
    "feedparser>=6.0",
    "trafilatura>=1.0",
    "python-dotenv>=1.0",
    "anthropic>=0.40",
    "openai>=1.0",
    "aiosqlite>=0.20",
]

[project.optional-dependencies]
audio = [
    "pydub>=0.25",
]
agent = [
    "python-telegram-bot>=20.0",
]
scheduler = [
    "apscheduler>=3.10",
]
all = [
    "nexus[audio,agent,web,scheduler]",
]
web = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "jinja2>=3.1",
    "markdown>=3.6",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nexus"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
